Add conftest for better organisation
